requirements.txt
# AI и обработка текста
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0
ollama>=0.1.6
groq>=0.4.0

# Парсинг веб-сайтов
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.1.0
selenium>=4.17.0
trafilatura>=1.6.3

# Работа с данными
pandas>=2.2.0
openpyxl>=3.1.2

# Утилиты
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0

# CLI
click>=8.1.7
rich>=13.7.0

# Работа с Google Ads API и MCP Server
google-ads>=23.1.0
PyYAML>=6.0.1

